+++ app/Http/Controllers/Import/Nordigen/LinkController.php
@@ -26,6 +26,7 @@ declare(strict_types=1);
 namespace App\Http\Controllers\Import\Nordigen;
 
 use App\Exceptions\ImporterErrorException;
+use App\Exceptions\ImporterHttpException;
 use App\Http\Controllers\Controller;
 use App\Http\Middleware\LinkControllerMiddleware;
 use App\Services\Nordigen\Request\GetRequisitionRequest;
@@ -90,9 +91,7 @@ class LinkController extends Controller
             return redirect(route('004-configure.index'));
         }
 
-        // create and save local reference:
-        $uuid = Uuid::uuid4()->toString();
-
+        $uuid        = (string) Uuid::uuid4()->toString();
         $url         = config('nordigen.url');
         $accessToken = TokenManager::getAccessToken();
         $request     = new PostNewRequisitionRequest($url, $accessToken);
@@ -100,12 +99,12 @@ class LinkController extends Controller
         $request->setBank($configuration->getNordigenBank());
         $request->setReference($uuid);
 
-        app('log')->debug(sprintf('Reference is %s', $uuid));
+        app('log')->debug(sprintf('Reference is "%s"', $uuid));
 
         /** @var NewRequisitionResponse $response */
         $response = $request->post();
-        app('log')->debug(sprintf('Got a new requisition with id %s', $response->id));
-        app('log')->debug(sprintf('Status: %s, returned reference: %s', $response->status, $response->reference));
+        app('log')->debug(sprintf('Got a new requisition with id "%s"', $response->id));
+        app('log')->debug(sprintf('Status: %s, returned reference: "%s"', $response->status, $response->reference));
         app('log')->debug(sprintf('Will now redirect the user to %s', $response->link));
 
         // save config!
@@ -121,14 +120,19 @@ class LinkController extends Controller
      * @param Request $request
      * @return Application|RedirectResponse|Redirector
      * @throws ImporterErrorException
+     * @throws ImporterHttpException
      * @throws ContainerExceptionInterface
      * @throws NotFoundExceptionInterface
      */
     public function callback(Request $request)
     {
-        $reference = $request->get('ref');
+        $reference = (string) $request->get('ref');
         app('log')->debug(sprintf('Now at %s', __METHOD__));
-        app('log')->debug(sprintf('Reference is %s', $reference));
+        app('log')->debug(sprintf('Reference is "%s"', $reference));
+
+        if('' === $reference) {
+            throw new ImporterHttpException('The reference returned by Nordigen was unexpectedly empty.');
+        }
 
         // create a new config thing
         $configuration = $this->restoreConfiguration();